requirements.txt
streamlit==1.53.1
torch==2.9.1
torchvision==0.24.1
timm==1.0.22
albumentations==1.4.3
opencv-python-headless==4.12.0.88
numpy==2.2.6
pillow==12.0.0
